Fixed profiler daemon log extraction
