+++ redisbench_admin/profilers/perf.py
@@ -10,7 +10,6 @@ import re
 import subprocess
 import time
 
-
 from redisbench_admin.profilers.pprof import (
     PPROF_FORMAT_TEXT,
     run_pprof,
@@ -18,7 +17,6 @@ from redisbench_admin.profilers.pprof import (
 )
 from redisbench_admin.profilers.profilers import STACKCOLLAPSE_PATH, FLAMEGRAPH_PATH
 from redisbench_admin.utils.utils import whereis
-
 
 PERF_CALLGRAPH_MODE_DEFAULT = "fp"
 LINUX_PERF_SETTINGS_MESSAGE = (
