+++ redisbench_admin/run_remote/remote_failures.py
@@ -10,34 +10,34 @@ from redisbench_admin.utils.utils import upload_artifacts_to_s3
 
 
 def failed_remote_run_artifact_store(
-    args,
+    upload_results_s3,
     client_public_ip,
     dirname,
-    full_logfile,
-    logname,
+    remote_file,
+    local_file,
     s3_bucket_name,
     s3_bucket_path,
     username,
     private_key,
 ):
-    local_logfile = "{}/{}".format(dirname, logname)
+    local_file_fullpath = "{}/{}".format(dirname, local_file)
     logging.error(
-        "The benchmark returned an error exit status. Fetching remote logfile {} into {}".format(
-            full_logfile, local_logfile
+        "The benchmark returned an error exit status. Fetching remote file {} into {}".format(
+            remote_file, local_file_fullpath
         )
     )
     fetch_file_from_remote_setup(
         client_public_ip,
         username,
         private_key,
-        local_logfile,
-        full_logfile,
+        local_file_fullpath,
+        remote_file,
     )
-    if args.upload_results_s3:
+    if upload_results_s3:
         logging.info(
-            "Uploading logfile {} to s3. s3 bucket name: {}. s3 bucket path: {}".format(
-                local_logfile, s3_bucket_name, s3_bucket_path
+            "Uploading file {} to s3. s3 bucket name: {}. s3 bucket path: {}".format(
+                local_file_fullpath, s3_bucket_name, s3_bucket_path
             )
         )
-        artifacts = [local_logfile]
+        artifacts = [local_file_fullpath]
         upload_artifacts_to_s3(artifacts, s3_bucket_name, s3_bucket_path)
